Moved the Rating Breakdown Screenshot from Profile Panel scene to  Rating Breakdown Modal
